requirements.txt
behave
requests
websocket-client
jsonschema
python-dotenv
pytest-html
allure-behave
allure-pytest
pytest     